examples/fastapi-netflix-clone/requirements.txt
fastapi==0.73.0
uvicorn==0.17.0.post1
jinja2==3.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
